third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-5-expected.txt

Verify javascript outline


Running: testComputedProperty
Text:
    a.b[c] = function() { };
Outline:

Running: testComputedPropertyInTheMiddle
Text:
    a.b[c].d = function() { };
Outline:
    d():0:0

Running: testComputedPropertyForExpression
Text:
    (a || b).c = function() { };
Outline:
    c():0:0

Running: testPropertyStringLiteral
Text:
    var foo = { "bar": function() { }}
Outline:

Running: testPropertyNumberLiteral
Text:
    var foo = { 42: function() { }}
Outline:
